+++ sdk/webpubsub/Azure.Messaging.WebPubSub.Chat/tests/ChatRolesAndPermissionsTests.cs
@@ -11,56 +11,29 @@ namespace Azure.Messaging.WebPubSub.Chat.Tests
     public class ChatRolesAndPermissionsTests
     {
         [Test]
-        public void BuiltInUserPermissionValues()
-        {
-            Assert.That(UserPermissions.CreateRoom, Is.EqualTo("user.create_room"));
-            Assert.That(UserPermissions.FetchAllRooms, Is.EqualTo("user.fetch_all_rooms"));
-        }
-
-        [Test]
-        public void BuiltInRoomPermissionValues()
-        {
-            Assert.That(RoomPermissions.InviteUser, Is.EqualTo("room.invite"));
-            Assert.That(RoomPermissions.RemoveUser, Is.EqualTo("room.remove_user"));
-            Assert.That(RoomPermissions.History, Is.EqualTo("room.history"));
-            Assert.That(RoomPermissions.PublishMessage, Is.EqualTo("room.publish_message"));
-        }
-
-        [Test]
-        public void BuiltInRoleValues()
-        {
-            Assert.That(ChatRoles.UserNormal, Is.EqualTo("user.normal"));
-            Assert.That(ChatRoles.RoomMember, Is.EqualTo("room.member"));
-            Assert.That(ChatRoles.RoomOperator, Is.EqualTo("room.operator"));
-        }
-
-        [Test]
         public void CanBuildCustomRoleFromBuiltInPermissions()
         {
             var role = new WebPubSubChatRole(new[]
             {
-                RoomPermissions.PublishMessage,
-                RoomPermissions.History,
-                RoomPermissions.InviteUser,
+                ChatPermission.RoomPublishMessage,
+                ChatPermission.RoomHistory,
+                ChatPermission.RoomInvite,
             });
 
             Assert.That(
                 role.Permissions,
-                Is.EquivalentTo(new[] { "room.publish_message", "room.history", "room.invite" }));
+                Is.EquivalentTo(new[]
+                {
+                    ChatPermission.RoomPublishMessage,
+                    ChatPermission.RoomHistory,
+                    ChatPermission.RoomInvite,
+                }));
         }
 
         [Test]
         public void ChatRoleConstructor_NullPermissions_Throws()
         {
-            Assert.Throws<ArgumentNullException>(() => new WebPubSubChatRole((IEnumerable<string>)null));
-        }
-
-        [Test]
-        public void BuiltInRoleCanBeAssignedToRoomMember()
-        {
-            var member = new WebPubSubChatRoomMember(ChatRoles.RoomOperator);
-
-            Assert.That(member.RoleName, Is.EqualTo("room.operator"));
+            Assert.Throws<ArgumentNullException>(() => new WebPubSubChatRole((IEnumerable<ChatPermission>)null));
         }
     }
 }